+++ pyproject.toml
@@ -234,6 +234,7 @@ omit = [
     "src/maps/*",
     "src/export/const_definitions_exporter.py",
     "src/export/org_alarm_event_exporter.py",
+    "src/export/site_config_exporter.py",
     "src/firmware/firmware_manager.py",
     "src/site/bulk_radius_wlan_config_manager.py",
     "src/ui/tui.py",
